requirements.txt
streamlit
google-genai
google-api-python-client
feedparser
requests
beautifulsoup4